team10/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-WTF==0.14.2
idna==2.5
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
newsapi-python==0.2.3
requests==2.17.1
urllib3==1.21.1
Werkzeug==0.15.2
WTForms==2.2.1
